Create helper Document object
The Document object provides an API to modify a document to be sent. This is
useful for giving people a structured way to develop a document. It should
also hide some of the messy detials of JSON requests, example entering a
complete URI instead of just a name/id. Method chaining is available.